[PW-7028] - Move the voucher information box to the success page
