+++ src/Resources/app/storefront/src/finish/adyen-success-action.plugin.js
@@ -1,0 +1,47 @@
+/*
+ *                       ######
+ *                       ######
+ * ############    ####( ######  #####. ######  ############   ############
+ * #############  #####( ######  #####. ######  #############  #############
+ *        ######  #####( ######  #####. ######  #####  ######  #####  ######
+ * ###### ######  #####( ######  #####. ######  #####  #####   #####  ######
+ * ###### ######  #####( ######  #####. ######  #####          #####  ######
+ * #############  #############  #############  #############  #####  ######
+ *  ############   ############  #############   ############  #####  ######
+ *                                      ######
+ *                               #############
+ *                               ############
+ *
+ * Adyen Payment Module
+ *
+ * Copyright (c) 2022 Adyen N.V.
+ * This file is open source and available under the MIT license.
+ * See the LICENSE file for more info.
+ *
+ */
+
+import Plugin from 'src/plugin-system/plugin.class';
+import DomAccess from 'src/helper/dom-access.helper';
+import StoreApiClient from 'src/service/store-api-client.service';
+import ElementLoadingIndicatorUtil from 'src/utility/loading-indicator/element-loading-indicator.util';
+
+export default class AdyenSuccessActionPlugin extends Plugin {
+    init() {
+        this._client = new StoreApiClient();
+        this.adyenCheckout = Promise;
+        this.initializeCheckoutComponent().bind(this);
+    }
+
+    async initializeCheckoutComponent () {
+        const { locale, clientKey, environment, action} = adyenSuccessActionConfiguration;
+
+        const ADYEN_CHECKOUT_CONFIG = {
+            locale,
+            clientKey,
+            environment
+        };
+
+        this.adyenCheckout = await AdyenCheckout(ADYEN_CHECKOUT_CONFIG);
+        this.adyenCheckout.createFromAction(JSON.parse(action)).mount('#success-action-container');
+    }
+}
