+++ src/Resources/app/storefront/src/scss/base.scss
@@ -34,7 +34,8 @@
     vertical-align: top;
 }
 
-#donation-container {
+#donation-container,
+#success-action-container {
     margin-top: -50px;
     margin-bottom: 50px;
 }
