+++ src/Resources/config/services.xml
@@ -86,7 +86,7 @@
             <argument key="$logger" type="service" id="monolog.logger.adyen_generic"/>
             <tag name="kernel.event_subscriber"/>
         </service>
-        <service id="Adyen\Shopware\Subscriber\AdyenGivingSubscriber">
+        <service id="Adyen\Shopware\Subscriber\PostPaymentSubscriber">
             <argument type="service" id="Adyen\Shopware\Service\Repository\SalesChannelRepository"/>
             <argument type="service" id="Adyen\Shopware\Service\ConfigurationService"/>
             <argument type="service" id="Adyen\Util\Currency"/>
