+++ src/Resources/views/storefront/component/adyen-giving/adyen-giving-component.html.twig
@@ -1,6 +1,6 @@
-{% set adyenFrontendData = page.extensions[constant('Adyen\\Shopware\\Subscriber\\AdyenGivingSubscriber::ADYEN_DATA_EXTENSION_ID')] %}
+{% set adyenFrontendData = page.extensions[constant('Adyen\\Shopware\\Subscriber\\PostPaymentSubscriber::ADYEN_DATA_EXTENSION_ID')] %}
 
-{% if adyenFrontendData and adyenFrontendData.locale and adyenFrontendData.clientKey and adyenFrontendData.environment %}
+{% if adyenFrontendData and adyenFrontendData.locale and adyenFrontendData.clientKey and adyenFrontendData.environment and adyenFrontendData.givingEnabled %}
     <div id="adyen-giving-container">
         <div id="adyen-giving-configuration"
              data-locale="{{ adyenFrontendData.locale }}"
@@ -20,8 +20,6 @@
 
         <div id='donation-container'></div>
 
-        <link rel="stylesheet" href="{{ asset('bundles/adyenpaymentshopware6/css/adyen.css', 'asset') }}">
-        <script type="text/javascript" src="{{ asset('bundles/adyenpaymentshopware6/js/adyen.js', 'asset') }}"></script>
         <script>
             var adyenGivingConfiguration = document.querySelector('#adyen-giving-configuration').dataset;
         </script>
